EDITOR-#378 Add editingLEDEffect

+++ editor-server/src/resolvers/response/ledEffectResponse.ts
@@ -25,7 +25,7 @@ export class LEDEffectResponse {
 
 @ObjectType()
 class LEDEffects {
-  @Field((type) => Float)
+  @Field((type) => Int)
   start: number;
 
   @Field()
